# server/pkg/api/rest/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Collecting and Aggregating server module",
        "title": "CM-Honeybee Server REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/honeybee",
    "paths": {
...
        "/connection_info/{connId}": {
            "get": {
                "description": "Get the connection information directly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[On-premise] ConnectionInfo"
                ],
                "summary": "Get ConnectionInfo Directly",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the connectionInfo",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.ConnectionInfo"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/source_group/{sgId}/connection_info": {
            "get": {
                "description": "Get a list of connection information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[On-premise] ConnectionInfo"
                ],
                "summary": "List ConnectionInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the SourceGroup",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page of the connection information list.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Row of the connection information list.",
                        "name": "row",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name of the connection information.",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Description of the connection information.",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "IP address of the connection information.",
                        "name": "ip_address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SSH port of the connection information.",
                        "name": "ssh_port",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User of the connection information.",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get a list of connection information.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.ConnectionInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get a list of connection information.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create the connection information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[On-premise] ConnectionInfo"
                ],
                "summary": "Create ConnectionInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the SourceGroup",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Connection information of the node.",
                        "name": "ConnectionInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.CreateConnectionInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully register the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.ConnectionInfo"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to register the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/source_group/{sgId}/connection_info/{connId}": {
            "get": {
                "description": "Get the connection information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[On-premise] ConnectionInfo"
                ],
                "summary": "Get ConnectionInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the SourceGroup",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the connectionInfo",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.ConnectionInfo"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the connection information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[On-premise] ConnectionInfo"
                ],
                "summary": "Update ConnectionInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the SourceGroup",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the connectionInfo",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Connection information to modify.",
                        "name": "ConnectionInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.CreateConnectionInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.ConnectionInfo"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the connection information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[On-premise] ConnectionInfo"
                ],
                "summary": "Delete ConnectionInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the SourceGroup",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the connectionInfo",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SimpleMsg"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete the connection information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.ConnectionInfo": {
            "type": "object",
            "required": [
                "id",
                "ip_address",
                "name",
                "source_group_id",
                "ssh_port",
                "user"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "failed_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private_key": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "source_group_id": {
                    "type": "string"
                },
                "ssh_port": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.CreateConnectionInfoReq": {
            "type": "object",
            "required": [
                "ip_address",
                "name",
                "password",
                "ssh_port",
                "user"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private_key": {
                    "type": "string"
                },
                "ssh_port": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SimpleMsg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}